# api/swagger.yaml
---
swagger: "2.0"
info:
  description: "Simple API for the Corradini Center website."
  version: "1.1.0"
  title: "Corradini"
  contact:
    email: "gerlando.re@mail.polimi.it"
  license:
    name: "Corradini-Website APIs"
    url: "https://localhost:8080"
host: "localhost:8080"
basePath: "/crrdn"
tags:
- name: "role"
  description: "role within the association."
- name: "person"
  description: "a person in the association."
- name: "service"
  description: "a service offered by the association."
- name: "event"
  description: "an event organized by the association."
paths:
  /roles:
    get:
      tags:
      - "role"
      summary: "gets all the roles in the association"
      description: "This retrieves the group for free or all roles in the association"
      operationId: "rolesGET"
      produces:
      - "application/json"
      parameters: []
      responses:
        "200":
          description: "so a collection of roles"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Role"
        "404":
          description: "roles could not be found"
        default:
          description: "Something went wrong in the request."
      x-swagger-router-controller: "Role"
  /roles/{type}:
    get:
      tags:
      - "role"
      summary: "gets a group of roles of a given type, 'free' or 'all'"
      description: "retrieves all the roles of a given type, chosen by the corresponding\
        \ group"
      operationId: "rolesTypeGET"
      produces:
      - "application/json"
      parameters:
      - name: "type"
        in: "path"
        description: "type of group of role"
        required: true
        type: "string"
        enum:
        - "free"
        - "all"
      - name: "limit"
        in: "query"
        description: "max number of items per page"
        required: false
        type: "integer"
      - name: "offset"
        in: "query"
        description: "pagination offset for the given page"
        required: false
        type: "integer"
      responses:
        "200":
          description: "roles by category, so a collection of roles"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Role"
        "404":
          description: "roles could not be found"
        default:
          description: "Something went wrong in the request."
      x-swagger-router-controller: "Role"
  /roles/{type}/{rid}:
    get:
      tags:
      - "role"
      summary: "gets a specific role in a group of given type, specified by id"
      description: "retrieves a single role that is in a group of roles of a given\
        \ 'type'"
      operationId: "rolesTypeRidGET"
      produces:
      - "application/json"
      parameters:
      - name: "rid"
        in: "path"
        description: "pk of the role in the group of given type"
        required: true
        type: "string"
      - name: "type"
        in: "path"
        description: "type of group of role"
        required: true
        type: "string"
        enum:
        - "free"
        - "all"
      - name: "limit"
        in: "query"
        description: "max number of items per page"
        required: false
        type: "integer"
      - name: "offset"
        in: "query"
        description: "pagination offset for the given page"
        required: false
        type: "integer"
      responses:
        "200":
          description: "roles by category, so a collection of roles"
          schema:
            type: "object"
            maximum: 1
        "400":
          description: "given selector does not fit"
        "404":
          description: "roles could not be found"
        default:
          description: "Something went wrong in the request"
      x-swagger-router-controller: "Role"
  /roles/{rid}/relatedTo:
    get:
      tags:
      - "role"
      summary: "gets all the services that require the role with the given {rid}"
      description: "returns an array of services related to the role."
      operationId: "rolesRidRelatedToGET"
      produces:
      - "application/json"
      parameters:
      - name: "rid"
        in: "path"
        description: "pk of the role in the group of given type"
        required: true
        type: "string"
      - name: "limit"
        in: "query"
        description: "max number of items per page"
        required: false
        type: "integer"
      - name: "offset"
        in: "query"
        description: "pagination offset for the given page"
        required: false
        type: "integer"
      responses:
        "200":
          description: "array of services related to that role."
          schema:
            minimum: 1
            maximum: 10
            type: "array"
            items:
              $ref: "#/definitions/Service"
        "400":
          description: "given selector does not fit"
        "404":
          description: "services could not be found"
        default:
          description: "Something went wrong in the request"
      x-swagger-router-controller: "Role"
  /services:
    get:
      tags:
      - "service"
      summary: "gets all the services arranged by the association"
      description: "Take the list of all services in the association"
      operationId: "servicesGET"
      produces:
      - "application/json"
      parameters: []
      responses:
        "200":
          description: "service by category, so a collection of roles"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Service"
        "404":
          description: "services could not be found"
        default:
          description: "Something went wrong in the request"
      x-swagger-router-controller: "Service"
  /services/{type}:
    get:
      tags:
      - "service"
      summary: "gets all the services of a given type"
      description: "retrieves all the services of a given type."
      operationId: "servicesTypeGET"
      produces:
      - "application/json"
      parameters:
      - name: "type"
        in: "path"
        description: "type of group of role"
        required: true
        type: "string"
      - name: "limit"
        in: "query"
        description: "max number of items per page"
        required: false
        type: "integer"
      - name: "offset"
        in: "query"
        description: "pagination offset for the given page"
        required: false
        type: "integer"
      responses:
        "200":
          description: "services by category, so a collection of services"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Service"
        "404":
          description: "roles could not be found"
        default:
          description: "Something went wrong in the request"
      x-swagger-router-controller: "Service"
  /services/{type}/{sid}:
    get:
      tags:
      - "service"
      summary: "gets a specific service of a given type"
      description: "retrieves a service given the identifier and the type."
      operationId: "servicesTypeSidGET"
      produces:
      - "application/json"
      parameters:
      - name: "sid"
        in: "path"
        description: "identifier of the selected service"
        required: true
        type: "string"
      - name: "type"
        in: "path"
        description: "type of group of role"
        required: true
        type: "string"
      - name: "limit"
        in: "query"
        description: "max number of items per page"
        required: false
        type: "integer"
      - name: "offset"
        in: "query"
        description: "pagination offset for the given page"
        required: false
        type: "integer"
      responses:
        "200":
          description: "selected service by service identifier"
          schema:
            type: "object"
        "400":
          description: "given sid does not fit"
        "404":
          description: "roles could not be found"
        default:
          description: "Something went wrong in the request"
      x-swagger-router-controller: "Service"
  /services/{sid}/presentedIn:
    get:
      tags:
      - "service"
      summary: "gets a specific service of a given type"
      description: "retrieves all the events in which the service is presented."
      operationId: "servicesSidPresentedInGET"
      produces:
      - "application/json"
      parameters:
      - name: "sid"
        in: "path"
        description: "identifier of the selected service"
        required: true
        type: "string"
      - name: "limit"
        in: "query"
        description: "max number of items per page"
        required: false
        type: "integer"
      - name: "offset"
        in: "query"
        description: "pagination offset for the given page"
        required: false
        type: "integer"
      responses:
        "200":
          description: "events presenting the given service."
          schema:
            minimum: 0
            maximum: 10
            type: "array"
            items:
              $ref: "#/definitions/Event"
        "400":
          description: "given sid does not fit"
        "404":
          description: "events could not be found"
        default:
          description: "Something went wrong in the request"
      x-swagger-router-controller: "Service"
  /services/{sid}/involving:
    get:
      tags:
      - "service"
      summary: "gets all the people involved in a given Service"
      description: "retrieves the people helping in the service."
      operationId: "servicesTypeSidInvolvingGET"
      produces:
      - "application/json"
      parameters:
      - name: "sid"
        in: "path"
        description: "identifier of the selected service"
        required: true
        type: "string"
      - name: "limit"
        in: "query"
        description: "max number of items per page"
        required: false
        type: "integer"
      - name: "offset"
        in: "query"
        description: "pagination offset for the given page"
        required: false
        type: "integer"
      responses:
        "200":
          description: "people working for the selected service."
          schema:
            minimum: 1
            maximum: 5
            type: "array"
            items:
              $ref: "#/definitions/Person"
        "400":
          description: "given sid does not fit"
        "404":
          description: "people could not be found"
        default:
          description: "Something went wrong in the request"
      x-swagger-router-controller: "Service"
  /events:
    get:
      tags:
      - "event"
      summary: "gets all the group of events"
      description: "Take the list of every event"
      operationId: "eventsGET"
      produces:
      - "application/json"
      parameters: []
      responses:
        "200":
          description: "events by category, so a collection of group of events"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Event"
        "404":
          description: "events could not be found"
        default:
          description: "Something went wrong in the request"
      x-swagger-router-controller: "Event"
  /events/{month}:
    get:
      tags:
      - "event"
      summary: "gets all the events by the selected month"
      description: "retrieves all the events happening on a given month"
      operationId: "eventsMonthGET"
      produces:
      - "application/json"
      parameters:
      - name: "month"
        in: "path"
        description: "month in which there is all the group of events"
        required: true
        type: "integer"
      - name: "limit"
        in: "query"
        description: "max number of items per page"
        required: false
        type: "integer"
      - name: "offset"
        in: "query"
        description: "pagination offset for the given page"
        required: false
        type: "integer"
      responses:
        "200":
          description: "collection of events"
          schema:
            minimum: 0
            maximum: 10
            type: "array"
            items:
              $ref: "#/definitions/Event"
        "404":
          description: "roles could not be found"
        default:
          description: "Something went wrong in the request"
      x-swagger-router-controller: "Event"
  /events/{month}/{eid}:
    get:
      tags:
      - "event"
      summary: "gets an the event from the selected month"
      description: "retrieves the event in a given month with a given identifier."
      operationId: "eventsMonthEidGET"
      produces:
      - "application/json"
      parameters:
      - name: "eid"
        in: "path"
        description: "identifier for the event"
        required: true
        type: "string"
      - name: "month"
        in: "path"
        description: "month in which there is all the group of events"
        required: true
        type: "integer"
      - name: "limit"
        in: "query"
        description: "max number of items per page"
        required: false
        type: "integer"
      - name: "offset"
        in: "query"
        description: "pagination offset for the given page"
        required: false
        type: "integer"
      responses:
        "200":
          description: "event with specified id."
          schema:
            type: "object"
        "404":
          description: "roles could not be found"
        default:
          description: "Something went wrong in the request"
      x-swagger-router-controller: "Event"
  /events/{eid}/presents:
    get:
      tags:
      - "event"
      summary: "gets the services presented in the event"
      description: "retrieves all the services presented in the event specified by\
        \ the eid."
      operationId: "eventsEidPresentsGET"
      produces:
      - "application/json"
      parameters:
      - name: "eid"
        in: "path"
        description: "identifier for the event"
        required: true
        type: "string"
      - name: "limit"
        in: "query"
        description: "max number of items per page"
        required: false
        type: "integer"
      - name: "offset"
        in: "query"
        description: "pagination offset for the given page"
        required: false
        type: "integer"
      responses:
        "200":
          description: "roles by category, so a collection of roles"
          schema:
            minimum: 0
            maximum: 2
            type: "array"
            items:
              $ref: "#/definitions/Service"
        "404":
          description: "roles could not be found"
        default:
          description: "Something went wrong in the request"
      x-swagger-router-controller: "Event"
  /events/{eid}/contact:
    get:
      tags:
      - "event"
      summary: "gets the person in charge as contact for the event"
      description: "retrieves the person that is the official contact for that event."
      operationId: "eventsEidContactGET"
      produces:
      - "application/json"
      parameters:
      - name: "eid"
        in: "path"
        description: "identifier for the event"
        required: true
        type: "string"
      - name: "limit"
        in: "query"
        description: "max number of items per page"
        required: false
        type: "integer"
      - name: "offset"
        in: "query"
        description: "pagination offset for the given page"
        required: false
        type: "integer"
      responses:
        "200":
          description: "a person object"
          schema:
            type: "object"
            minimum: 1
            maximum: 1
        "404":
          description: "person could not be found"
        default:
          description: "Something went wrong in the request"
      x-swagger-router-controller: "Event"
  /people:
    get:
      tags:
      - "person"
      summary: "gets the list of all people in the association."
      description: "Take the list of all people"
      operationId: "peopleGET"
      produces:
      - "application/json"
      parameters: []
      responses:
        "200":
          description: "events by category, so a collection of group of events"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Person"
        "404":
          description: "people could not be found"
        default:
          description: "Something went wrong in the request"
      x-swagger-router-controller: "Person"
  /people/{job}:
    get:
      tags:
      - "person"
      summary: "gets all the people involved in a given type of job"
      description: "retrieves all the people having a given type of job in the association."
      operationId: "peopleJobGET"
      produces:
      - "application/json"
      parameters:
      - name: "job"
        in: "path"
        description: "role of the person"
        required: true
        type: "string"
      - name: "limit"
        in: "query"
        description: "max number of items per page"
        required: false
        type: "integer"
      - name: "offset"
        in: "query"
        description: "pagination offset for the given page"
        required: false
        type: "integer"
      responses:
        "200":
          description: "people by type of job"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Person"
        "404":
          description: "roles could not be found"
        default:
          description: "Something went wrong in the request"
      x-swagger-router-controller: "Person"
  /people/{job}/{pid}:
    get:
      tags:
      - "person"
      summary: "gets a specific person that has a given job"
      description: "retrieves a person with a specific id and a type of job"
      operationId: "peopleJobPidGET"
      produces:
      - "application/json"
      parameters:
      - name: "pid"
        in: "path"
        description: "role of the person"
        required: true
        type: "string"
      - name: "job"
        in: "path"
        description: "role of the person"
        required: true
        type: "string"
      - name: "limit"
        in: "query"
        description: "max number of items per page"
        required: false
        type: "integer"
      - name: "offset"
        in: "query"
        description: "pagination offset for the given page"
        required: false
        type: "integer"
      responses:
        "200":
          description: "person"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Person"
        "404":
          description: "roles could not be found"
        default:
          description: "Something went wrong in the request"
      x-swagger-router-controller: "Person"
  /people/{pid}/role:
    get:
      tags:
      - "person"
      summary: "retrieves the role of a specific person"
      description: "retrieves the role of a specific person"
      operationId: "peoplePidRoleGET"
      produces:
      - "application/json"
      parameters:
      - name: "pid"
        in: "path"
        description: "role of the person"
        required: true
        type: "string"
      responses:
        "200":
          description: "role of the selected person"
          schema:
            type: "object"
            minimum: 1
            maximum: 1
        "404":
          description: "role could not be found"
        default:
          description: "Something went wrong in the request"
      x-swagger-router-controller: "Person"
  /people/{pid}/contactFor:
    get:
      tags:
      - "person"
      summary: "gets a specific person that has a given job"
      description: "retrieves the events for which the person is the contact."
      operationId: "peoplePidContactForGET"
      produces:
      - "application/json"
      parameters:
      - name: "pid"
        in: "path"
        description: "role of the person"
        required: true
        type: "string"
      - name: "limit"
        in: "query"
        description: "max number of items per page"
        required: false
        type: "integer"
      - name: "offset"
        in: "query"
        description: "pagination offset for the given page"
        required: false
        type: "integer"
      responses:
        "200":
          description: "events for which the person is the contact"
          schema:
            minimum: 0
            maximum: 4
            type: "array"
            items:
              $ref: "#/definitions/Event"
        "404":
          description: "events could not be found"
        default:
          description: "Something went wrong in the request"
      x-swagger-router-controller: "Person"
  /people/{pid}/involved:
    get:
      tags:
      - "person"
      summary: "gets the services in which the person works."
      description: "retrieves the services in which the person works."
      operationId: "peoplePidInvolvedGET"
      produces:
      - "application/json"
      parameters:
      - name: "pid"
        in: "path"
        description: "identifier of the person"
        required: true
        type: "string"
      - name: "limit"
        in: "query"
        description: "max number of items per page"
        required: false
        type: "integer"
      - name: "offset"
        in: "query"
        description: "pagination offset for the given page"
        required: false
        type: "integer"
      responses:
        "200":
          description: "services for which the person is the contact"
          schema:
            minimum: 1
            maximum: 3
            type: "array"
            items:
              $ref: "#/definitions/Service"
        "404":
          description: "services could not be found"
        default:
          description: "Something went wrong in the request"
      x-swagger-router-controller: "Person"
definitions:
  Role:
    type: "object"
    required:
    - "description"
    - "rid"
    - "type"
    properties:
      rid:
        type: "string"
        minLength: 4
        maxLength: 4
      rolename:
        type: "string"
      description:
        type: "string"
      type:
        type: "string"
      freeroles:
        type: "integer"
    example:
      freeroles: 0
      rolename: "rolename"
      description: "description"
      rid: "rid"
      type: "type"
  Service:
    type: "object"
    required:
    - "description"
    - "place"
    - "sid"
    - "type"
    properties:
      sid:
        type: "string"
        minLength: 4
        maxLength: 4
      servicename:
        type: "string"
      description:
        type: "string"
      type:
        type: "string"
      place:
        type: "string"
    example:
      description: "description"
      servicename: "servicename"
      place: "place"
      type: "type"
      sid: "sid"
  Event:
    type: "object"
    required:
    - "contact"
    - "description"
    - "eid"
    properties:
      eid:
        type: "string"
        minLength: 4
        maxLength: 4
      eventname:
        type: "string"
      description:
        type: "string"
      begintime:
        type: "string"
        format: "date-time"
      endtime:
        type: "string"
        format: "date-time"
      schedule:
        type: "integer"
      place:
        type: "string"
      contact:
        type: "string"
        minLength: 4
        maxLength: 4
    example:
      eid: "eid"
      schedule: 0
      contact: "contact"
      description: "description"
      endtime: "2000-01-23T04:56:07.000+00:00"
      begintime: "2000-01-23T04:56:07.000+00:00"
      place: "place"
      eventname: "eventname"
  Person:
    type: "object"
    required:
    - "email"
    - "pid"
    properties:
      pid:
        type: "string"
        minLength: 4
        maxLength: 4
      firstname:
        type: "string"
      lastname:
        type: "string"
      email:
        type: "string"
        format: "email"
      phonenumber:
        type: "string"
      age:
        type: "integer"
        minimum: 0
        maximum: 99
      role:
        type: "string"
        minLength: 4
        maxLength: 4
    example:
      firstname: "firstname"
      role: "role"
      phonenumber: "phonenumber"
      pid: "pid"
      email: "email"
      age: 7
      lastname: "lastname"